18/05/2021

+++ main.c
@@ -4,8 +4,9 @@
 *                                    FPII                                       *
 ********************************************************************************/
 
-#include <stdio.h> 
-#include <stdbool.h> 
+#include <stdio.h>
+#include <stdbool.h>
+#include <stdlib.h>
 #include <time.h>
 
 /* Incluimos el fichero de las cabezeras de los procedimientos */
@@ -13,14 +14,14 @@
 
 
 /** Programa principal **/
-int main() 
+int main()
 {
 	/* Declaracion de variables */
 	int opcion = 0;
 	bool salir = false; /* Booleano para salir del programa principal, y repetir el menu */
 	int vector[MAX_TABLA], num_columnas, contador; /* Tabla cargada con la que se trabaja */
 
-	srand((unsigned int)time(NULL)); // Iniciamos la semilla para generar numeros aleatorios
+	srand(time(NULL)); // Iniciamos la semilla para generar numeros aleatorios
 
 	/* Bucle del Menu y los procedimientos */
 	while (!salir)
@@ -41,17 +42,17 @@ int main()
 			scanf("%d", &opcion);
 			printf("\n ------------------------------------\n\n");
 		} while (opcion < 0 || opcion > 6);
-		
+
         /* Realizamos las opciones */
 		switch (opcion)
-		{	
+		{
 			case 0: /** SALIR **/
                 salir = true;
 				printf(" Gracias por utilizar el programa\n");
 				printf("\n ------------------------------------\n");
 				break;
-				
-			
+
+
 			case 1: /** Ordenar por seleccion **/
 				crear_vector(vector, &num_columnas);
 				ver_vector(vector, num_columnas);
@@ -62,7 +63,7 @@ int main()
 				printf(" Se han realizado %d intercambios\n", contador);
                 printf("\n ------------------------------------\n");
 				break;
-			
+
 			case 2: /** Ordenar por burbuja **/
 				crear_vector(vector, &num_columnas);
 				ver_vector(vector, num_columnas);
@@ -73,7 +74,7 @@ int main()
 				printf(" Se han realizado %d intercambios\n", contador);
                 printf("\n ------------------------------------\n");
 				break;
-			
+
 			case 3: /** Ordenar por insercion **/
 				crear_vector(vector, &num_columnas);
 				ver_vector(vector, num_columnas);
@@ -84,7 +85,7 @@ int main()
 				printf(" Se han realizado %d intercambios\n", contador);
                 printf("\n ------------------------------------\n");
 				break;
-			
+
 			case 4: /** Ver contenido del vector **/
 				crear_vector(vector, &num_columnas);
 				orden_seleccion(vector, num_columnas);
@@ -93,18 +94,18 @@ int main()
 
                 printf("\n ------------------------------------\n");
 				break;
-			
+
 			case 5: /** Prueba sistemas de ordenacion **/
 				test_ordenacion(vector, &num_columnas);
 				test_caso_mejor(vector, &num_columnas);
                 printf("\n ------------------------------------\n");
 				break;
-			
+
 			case 6: /** Prueba busqueda dicotomica **/
 				test_busqueda(vector, &num_columnas);
                 printf("\n ------------------------------------\n");
 				break;
 		}
 	}
-	return 0; 
-} 
+	return 0;
+}
